add security oauth2 ok

+++ doc/sql/account.sql
@@ -13,7 +13,7 @@ CREATE TABLE `user` (
     `email` VARCHAR(255) NULL,
     `phone` VARCHAR(255) NULL,
     `location` VARCHAR(255) NULL,
-    `archived` INT(1) NOT NULL DEFAULT 0,
+    `archived` TINYINT(1) NOT NULL DEFAULT 0,
     `created_at` DATETIME NOT NULL COMMENT '创建时间',
     `update_at` DATETIME NOT NULL COMMENT '信息更新时间',
     `last_login_at` DATETIME NOT NULL COMMENT '最后登录时间',
